requirements.txt
accelerate>=0.26.0
datasets
google.generativeai
huggingface_hub
ipykernel
ipywidgets
langchain
langchain-community
langchain_google_genai
matplotlib
numpy
opencv-python
pillow
python-dotenv
qwen-vl-utils
saliency
scikit-learn
sentencepiece
shapely
torch
torchvision
transformers
ultralytics